backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib
python-jose
bcrypt
motor
pydantic[email]
python-multipart
pydantic-settings
